feat: add gate check for delete subcategory

+++ backend/app/Http/Controllers/API/v1/Admin/Entity/SubCategory/SubCategoryController.php
@@ -102,6 +102,10 @@ class SubcategoryController extends Controller
      */
     public function destroy(Subcategory $subCategory): JsonResponse
     {
+        if(Gate::denies('delete', $subCategory)) {
+            return response()->json(['message' => 'You are not allowed to access this action'], 403);
+        }
+
         return $this->service->destroy($subCategory);
     }
